requirements.txt
appnope==0.1.0
Authlib==0.13
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
blis==0.4.1
boto3==1.10.28
botocore==1.13.28
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cycler==0.10.0
cymem==2.0.3
decorator==4.4.1
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
Flask==1.1.1
gunicorn==20.0.4
idna==2.8
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
more-itertools==7.2.0
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
numpy==1.17.4
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
patsy==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
plac==1.1.3
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
Pygments==2.4.2
pyparsing==2.4.5
pyrsistent==0.15.5
PySimpleGUI==4.7.0
python-dateutil==2.8.0
python-dotenv==0.8.2
pytz==2019.3
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.22.0
s3transfer==0.2.1
scikit-learn==0.21.3
scipy==1.3.2
Send2Trash==1.5.0
six==1.12.0
spacy==2.2.2
srsly==0.2.0
statsmodels==0.10.1
terminado==0.8.3
testpath==0.4.4
thinc==7.3.1
tornado==6.0.3
tqdm==4.38.0
traitlets==4.3.3
urllib3==1.25.7
wasabi==0.4.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
zipp==0.6.0
